+++ frontend/src/Components/CustomerMG/Spinner.jsx
