+++ frontend/src/Components/CustomerMG/login/LoginHome.jsx
@@ -1,0 +1,112 @@
+import React, { useEffect, useState } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
+import axios from "axios";
+import { BiChat } from "react-icons/bi"; // Importing the chat icon
+import { motion } from "framer-motion"; // Importing Framer Motion for animations
+
+function LoginHome() {
+  const location = useLocation();
+  const navigate = useNavigate();
+  const [user, setUser] = useState(location.state?.user);
+  const [loading, setLoading] = useState(!user);
+
+  useEffect(() => {
+    if (!user) {
+      axios
+        .get("http://localhost:5579/user", { withCredentials: true })
+        .then((response) => {
+          if (response.data.user) {
+            setUser(response.data.user);
+          } else {
+            navigate("/login");
+          }
+        })
+        .catch(() => navigate("/login"))
+        .finally(() => setLoading(false));
+    } else {
+      setLoading(false);
+    }
+  }, [user, navigate]);
+
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center h-screen bg-[#FCFAEE]">
+        <h1 className="text-2xl font-bold text-[#252B42]">Loading...</h1>
+      </div>
+    );
+  }
+
+  return (
+    <div className="flex flex-col justify-center items-center h-screen bg-[#FCFAEE] relative overflow-hidden">
+      {/* Background Gradient */}
+      <div className="absolute inset-0 bg-gradient-to-br from-[#FCFAEE] to-[#98DED9] z-0 opacity-100"></div>
+
+      {/* Decorative Background Shapes */}
+      <div className="absolute top-0 left-0 w-1/3 h-1/3 rounded-full bg-[#98DED9] opacity-20 blur-3xl z-10 transform -translate-x-1/2 -translate-y-1/2"></div>
+      <div className="absolute bottom-0 right-0 w-1/4 h-1/4 rounded-full bg-[#98DED9] opacity-10 blur-2xl z-10 transform translate-x-1/2 translate-y-1/2"></div>
+
+      {/* Animated Welcome Message */}
+      <motion.p
+        className="text-[#252B42] text-5xl font-bold mb-8 drop-shadow-lg"
+        initial={{ y: -50, opacity: 0 }}
+        animate={{ y: 0, opacity: 1 }}
+        transition={{ duration: 0.5 }}
+      >
+        Welcome, {user && user.name}!
+      </motion.p>
+
+      <motion.div
+        className="relative bg-white shadow-2xl rounded-3xl p-10 text-center max-w-lg w-full transition-transform duration-300 transform hover:scale-105 hover:shadow-2xl z-10"
+        initial={{ scale: 0.9 }}
+        animate={{ scale: 1 }}
+        transition={{ duration: 0.5 }}
+      >
+        {/* Gradient Background Inside the Card */}
+        <div className="absolute inset-0 bg-gradient-to-r from-[#98DED9] to-[#FCFAEE] opacity-70 rounded-3xl z-[-1]"></div>
+
+        {user && user.role === "ADMIN" && (
+          <motion.div
+            onClick={() => navigate("/admin")}
+            className="cursor-pointer text-lg text-[#FFFFFF] hover:text-gray-800 transition-all duration-300 mb-4 p-3 rounded-lg bg-[#23A6F0] shadow-md hover:shadow-lg transform"
+            whileHover={{ scale: 1.02 }} // Subtle scale effect
+            whileTap={{ scale: 0.95 }} // Subtle tap effect
+          >
+            Manage Tickets
+          </motion.div>
+        )}
+
+        {user && user.role === "CUSTOMER" && (
+          <div className="flex flex-col items-center gap-6">
+            {/* My Tickets */}
+            <motion.div
+              onClick={() => navigate("/home")}
+              className="cursor-pointer text-lg text-[#FFFFFF] hover:text-gray-800 transition-all duration-300 mb-4 p-3 rounded-lg bg-[#23A6F0] shadow-md hover:shadow-lg transform"
+              whileHover={{ scale: 1.02 }} // Subtle scale effect
+              whileTap={{ scale: 0.95 }} // Subtle tap effect
+            >
+              My Tickets
+            </motion.div>
+
+            {/* Chatbot Navigation */}
+            <motion.div
+              onClick={() => navigate("/chatbot")}
+              className="cursor-pointer flex items-center justify-center gap-2 text-lg text-[#FFFFFF] hover:text-gray-800 transition-all duration-300 mb-4 p-3 rounded-lg bg-[#23A6F0] shadow-md hover:shadow-lg transform"
+              whileHover={{ scale: 1.02 }} // Subtle scale effect
+              whileTap={{ scale: 0.95 }} // Subtle tap effect
+            >
+              <BiChat className="text-3xl" />
+              Chat with ECo
+            </motion.div>
+          </div>
+        )}
+      </motion.div>
+
+      {/* Footer Section */}
+      <footer className="absolute bottom-0 w-full text-[#FFFFFF] text-sm mb-4 opacity-90 bg-[#252B42] p-3 rounded-md text-center">
+        <p>© 2024 Customer Support. All rights reserved.</p>
+      </footer>
+    </div>
+  );
+}
+
+export default LoginHome;
