--- frontend/src/Components/Customer_suport/ChatBot.css
@@ -1,199 +1,0 @@
-/* Global reset */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-size: 0.9rem; /* Global font size */
-}
-
-/* Chat container */
-.chat-container {
-  display: flex;
-  flex-direction: column;
-  height: 600px;
-  width: 100%;
-  max-width: 450px;
-  border-radius: 12px;
-  box-shadow: 0px 8px 25px rgba(0, 0, 0, 0.2);
-  overflow: hidden;
-  background: linear-gradient(135deg, #3b82f6, #1e40af);
-  font-family: 'Roboto', sans-serif;
-  margin: 50px auto;
-}
-
-/* Chat header */
-.chat-header {
-  background-color: #005bbb;
-  color: white;
-  text-align: center;
-  padding: 15px;
-  font-size: 1em;
-}
-
-.header-content {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chat-logo {
-  display: none; /* Hides the logo as requested */
-}
-
-/* Chat box */
-.chat-box {
-  flex: 1;
-  padding: 15px;
-  overflow-y: auto;
-  background-color: #f5f5f5;
-  border-top-left-radius: 12px;
-  border-top-right-radius: 12px;
-}
-
-/* Chat message */
-.chat-message {
-  display: flex;
-  align-items: flex-start;
-  margin-bottom: 15px;
-  animation: fadeIn 0.5s forwards;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-/* User message alignment */
-.chat-message.user {
-  justify-content: flex-end; /* Align user messages to the right */
-}
-
-.chat-message.user .avatar {
-  margin-left: 10px; /* Space between avatar and message */
-}
-
-/* Bot message alignment */
-.chat-message.bot {
-  justify-content: flex-start; /* Align bot messages to the left */
-}
-
-.avatar {
-  width: 30px; /* Adjusted avatar size */
-  height: 30px;
-  border-radius: 50%; /* Circular avatars */
-  margin-right: 10px;
-}
-
-.user-avatar {
-  background-color: #598cdf; /* Blue circle for user avatar */
-}
-
-.bot-avatar {
-  background-color: #cccccc; /* Gray circle for bot avatar */
-}
-
-/* Message styles */
-.message {
-  max-width: 70%; /* Adjusted width for better visibility */
-  padding: 15px 20px; /* Padding for the message box */
-  border-radius: 20px; /* Rounded corners for messages */
-  background-color: white; /* White background for messages */
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-}
-
-.message-text {
-  font-size: 0.9rem;
-  color: black; /* Text color */
-}
-
-.chat-message.user .message {
-  background-color: #3b83f6b4; /* User message background */
-  color: white; /* User message text color */
-}
-
-.chat-message.bot .message {
-  background-color: #e0e0e0; /* Bot message background */
-}
-
-/* Timestamp styles */
-.timestamp {
-  font-size: 0.75em;
-  color: #777; /* Timestamp text color */
-  margin-top: 5px;
-  display: block;
-}
-
-/* Input area */
-.input-area {
-  display: flex;
-  padding: 10px 20px;
-  background-color: white;
-  border-top: 1px solid #ddd;
-  align-items: center; /* Center the elements vertically */
-}
-
-.chat-input {
-  flex: 1; /* Makes the input take the remaining space */
-  padding: 10px;
-  font-size: 0.85rem;
-  border-radius: 20px;
-  border: 1px solid #ddd;
-  margin-right: 10px;
-  outline: none;
-  transition: border-color 0.3s;
-}
-
-.chat-input:focus {
-  border-color: #3b82f6; /* Border color on focus */
-}
-
-.send-button {
-  width: 20%;
-  padding: 5px 12px; /* Smaller padding to reduce button size */
-  border-radius: 20px;
-  background-color: #3b82f6; /* Button background color */
-  color: white; /* Button text color */
-  border: none;
-  font-size: 0.85rem; /* Smaller font size for button */
-  cursor: pointer;
-  transition: background-color 0.3s; /* Smooth background color transition */
-  flex-shrink: 0; /* Prevents the button from shrinking */
-}
-
-.send-button:hover {
-  background-color: #005bb5; /* Button hover color */
-}
-
-/* Typing indicator */
-.typing-indicator {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-.dot {
-  height: 10px;
-  width: 10px;
-  margin: 0 3px;
-  background-color: #3b82f6; /* Color of the dots */
-  border-radius: 50%; /* Circular dots */
-  animation: blink 1s infinite ease-in-out alternate; /* Blinking animation */
-}
-
-@keyframes blink {
-  from {
-    opacity: 0.2; /* Start with low opacity */
-  }
-  to {
-    opacity: 1; 
-  }
-}
