--- frontend/src/Components/Customer_suport/components/home/PDFViewer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Viewer } from "@react-pdf-viewer/core";
-import "@react-pdf-viewer/core/lib/styles/index.css"; // Import the default styles
-
-const PDFViewer = ({ url }) => {
-  return (
-    <div style={{ height: "750px" }}>
-      <Viewer fileUrl={url} />
-    </div>
-  );
-};
-
-export default PDFViewer;
