--- frontend/src/Components/UserReview_MG/css/AdminPanelReviews.css
@@ -1,76 +1,0 @@
-.admin-reviews-container {
-    max-width: 1000px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: white;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    text-align: center;
-}
-
-.admin-reviews-list {
-    margin-top: 20px;
-    overflow-x: auto;
-}
-
-.reviews-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 20px;
-    text-align: left;
-}
-
-.reviews-table th, .reviews-table td {
-    padding: 10px;
-    border: 1px solid #ddd;
-}
-
-.reviews-table th {
-    background-color: #f4f4f4;
-    color: #333;
-    font-weight: bold;
-}
-
-.reviews-table td {
-    color: #555;
-}
-
-.delete-button {
-    padding: 5px 10px;
-    background-color: red;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-.delete-button:hover {
-    background-color: darkred;
-}
-
-.search-box {
-    padding: 10px;
-    width: 300px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    margin-bottom: 20px;
-    margin-right: 20px;
-}
-
-.pdf-button {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin-bottom: 20px;
-}
-
-.pdf-button:hover {
-    background-color: #0056b3;
-}
-
-.chart-container {
-    margin-top: 30px;
-}
