--- frontend/src/Components/UserReview_MG/css/StarRating.css
@@ -1,15 +1,0 @@
-.star-rating {
-    display: inline-flex;
-    gap: 5px;
-  }
-  
-  .star {
-    color: lightgray;
-    cursor: pointer;
-    font-size: 1.5rem;
-  }
-  
-  .star.filled {
-    color: gold;
-  }
-
