--- frontend/src/Components/UserReview_MG/css/UserAddReviews.css
@@ -1,236 +1,0 @@
-body {
-    margin: 0;
-    padding: 0;
-    font-family: 'Roboto', sans-serif;
-    background-color: #f5f7fa;
-}
-
-.user-reviews-container {
-    max-width: 800px;
-    margin: 50px auto;
-    padding: 40px;
-    background-color: #ffffff;
-    border-radius: 10px;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-    text-align: center;
-    font-family: 'Roboto', sans-serif;
-}
-
-h2, h3 {
-    color: #333;
-    margin-bottom: 20px;
-    font-weight: 700;
-    font-size: 24px;
-    text-transform: uppercase;
-}
-
-
-.review-form {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin-bottom: 30px;
-}
-
-.review-form input,
-.review-form textarea {
-    width: 100%;
-    padding: 12px;
-    border: 1px solid #d1d1d1;
-    border-radius: 6px;
-    font-size: 16px;
-    transition: border-color 0.3s;
-}
-
-.review-form input:focus,
-.review-form textarea:focus {
-    border-color: #007bff;
-    outline: none;
-}
-
-.review-form textarea {
-    min-height: 120px;
-    resize: vertical;
-}
-
-.review-form button {
-    padding: 12px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 6px;
-    font-size: 16px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.review-form button:hover {
-    background-color: #0056b3;
-}
-
-
-.open-modal-button {
-    padding: 12px 25px;
-    background-color: #28a745;
-    color: white;
-    border: none;
-    border-radius: 6px;
-    font-size: 16px;
-    font-weight: 600;
-    margin-bottom: 30px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.open-modal-button:hover {
-    background-color: #218838;
-}
-
-.review-modal {
-    background-color: white;
-    padding: 40px;
-    border-radius: 10px;
-    max-width: 500px;
-    width: 100%;
-    margin: 0 auto;
-    position: relative;
-}
-
-.review-modal-overlay {
-    background-color: rgba(0, 0, 0, 0.7);
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.close-modal-button {
-    margin-top: 20px;
-    padding: 12px;
-    background-color: #dc3545;
-    color: white;
-    border: none;
-    border-radius: 6px;
-    font-size: 16px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.close-modal-button:hover {
-    background-color: #c82333;
-}
-
-.reviews-list {
-    margin-top: 30px;
-    text-align: left;
-}
-
-.review-item {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    padding: 20px;
-    border-radius: 6px;
-    border: 1px solid #e0e0e0;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
-    margin-bottom: 20px;
-    background-color: #f9f9f9;
-}
-
-.review-details {
-    flex: 1;
-    margin-right: 20px;
-}
-
-.review-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 10px;
-}
-
-.review-username {
-    color: #0056b3;
-    font-weight: bold;
-    margin: 0;
-    font-size: 18px;
-}
-
-.review-message {
-    color: #555;
-    line-height: 1.6;
-    margin-right: 450px;
-    font-size: 16px;
-}
-
-.star-icon {
-    color: #f39c12;
-    font-size: 20px;
-    margin-right: 2px;
-}
-
-.review-rating {
-    margin-top: 5px;
-    display: flex;
-    align-items: center;
-}
-
-.review-actions {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-.edit-icon,
-.delete-icon {
-    cursor: pointer;
-    font-size: 20px;
-    margin-left: 10px;
-    transition: color 0.3s ease;
-}
-
-.edit-icon:hover {
-    color: #28a745;
-}
-
-.delete-icon:hover {
-    color: #dc3545;
-}
-
-.rating-summary {
-    margin-top: 30px;
-    padding: 20px;
-    background-color: #f1f1f1;
-    border-radius: 8px;
-    text-align: left;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
-}
-
-.rating-summary div {
-    margin-bottom: 10px;
-}
-
-.rating-summary span {
-    font-size: 16px;
-    font-weight: 500;
-}
-
-@media (max-width: 768px) {
-    .user-reviews-container {
-        padding: 20px;
-    }
-
-    .review-item {
-        flex-direction: column;
-    }
-
-    .review-details {
-        margin-right: 0;
-    }
-}
